+++ src/test/java/crossover/methods/BasicOnePointCrossoverTest.java
@@ -43,12 +43,12 @@ public class BasicOnePointCrossoverTest {
     }
 
     private PersonTemplate preparePersonTemplate() {
-        List<Pair<Integer, Integer>> limits = Arrays.asList(
-                new Pair<>(0, 10),
-                new Pair<>(5, 7),
-                new Pair<>(10,20),
-                new Pair<>(100, 200),
-                new Pair<>(-20, -10)
+        List<Pair<Double, Double>> limits = Arrays.asList(
+                new Pair<>(0.0d, 10.0d),
+                new Pair<>(5.0d, 7.0d),
+                new Pair<>(10.0d,20.0d),
+                new Pair<>(100.0d, 200.0d),
+                new Pair<>(-20.0d, -10.0d)
         );
 
         PersonTemplate template = new PersonTemplate(SIZE, limits);
